pyproject.toml
[project]
name = "anime-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "onnxruntime<1.20",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
